refector: divide footer and header into different components
