+++ src/components/GlobalLayout/internal/constant.ts
